Add _AllIPs() config method
This returns a list of tuples (ip, network). Use it in ReserveIp,
where a network without AddressPool can be passed, in order to
ensure IP uniqueness inside the network. Please note that no
configuration error is raised if a two NICs have the same IP but do
not reside in a network.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>