Various refactoring concerning networks
Add external flag to UnlockedCommitIp.  Use this flag from cmdlib to
reserve and release IPs via TemporaryReservationManager and not
directly by accessing the pool it self.

Remove checks of style IsReserved() from cmdlib, while cfg.ReserveIp()
does this check and raises the corresponding error.

Reduce invocations of self.proc.GetEDId(). Store this in a
temporary var and use it inside a method.

Delegate exception handling in config/network.  Modify
ReserveIp/GenerateIP to raise OpPrereqError. So there is no need to
try-except in cmdlib.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>