Use new network interface
Replace old network.AddressPool() with network.Network().  Replace
network.AddressPool and network.AddressPool.InitializeNetwork with
network.Network. Additionally be aware that this raises
OpPrereqError and not AddressPoolError so not try-except-raise is
needed.

Make network optional in Network opcodes.  If no --network is passed
in OpNetworkAdd then a GenericNetwork is created without an IPv4
Address Pool. In this case no stats are available for the network so
client must format the output accordingly.

Currently removing/modifying a subnet will not be supported.

Now we can define solely collision domains (connected networks with
connectivity mode and link), IPv4 networks, IPv6 networks and all these
combined.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>