debian: Run autogen.sh before configure
This is needed so that packaging can be done by using official
src tree found on git repo.

Official debian packages provide in orig tarball the files
that get generated by automake, autoconf, and aclocal:

 * configure
 * Makefile.in
 * autotools scripts

We decide not to track those files in upstream branch (which are
already included in .gitignore), instead we generate those
files on every build.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>