kvm: Support QEMU logging
QEMU commits 96c33a4 and c586eac, in case of daemonize,
redirect stderr to the logfile, if given.

Prior to these commits the -D option was ignored and stderr
we redirected to /dev/null.

Here we pass the -D option along with the -daemonize option in the
QEMU command line. We use a dedicated directory for the QEMU logs,
and let QEMU use /var/log/ganeti/kvm/<instance name>.log to log its
stderr.

Signed-off-by: Dimitris Aragiorgis <dimara@arrikto.com>