backend: Use uuid and not idx for bdev symlinks
Until now, backend used the disk's index when creating the symlink
of the corresponding block device. This is problematic. For example
if one starts an instance with three disks [disk0, disk1, disk2],
then removes the middle one (disk1), and then adds a third (disk3),
the disks in config data will be [disk1, disk2, disk3] and thus the
new one will get index 2. When trying to assemble the newly created
disk we will overwrite the disk2's symlink. Fix the above behavior
by creating symlinks based on the disk's uuid and not its index.

To keep old instances migratable, we still create the old type
of symlink during BlockdevOpen() which is invoked on target node
just before migration.

Also remove a really old check that did not make any sense anymore
in GetInstanceMigratable().

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>