Fix failover in case the source node is offline
Until now failover did _CloseInstanceDisks() on the source node
before switching nodes. In case the node is offline this fails.
Since we might have data corruption (local caches), the user must
explicitly pass --ignore-consistency to proceed with the failover if
the node is unreachable or for some reason the RPC fails.

This fixes #1162.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>